make deletion of message optional

+++ src/main/java/net/nhalrath/DisSMA/listeners/WordFilter.java
@@ -47,10 +47,12 @@ public class WordFilter extends ListenerAdapter {
 
     private HashMap<String, Integer> memberOffenses;
     private List<String> badWords;
+    private boolean strictMessageFilter;
 
-    public WordFilter() {
-        badWords = new ArrayList<String>();
-        memberOffenses = new HashMap<String, Integer>();
+    public WordFilter(boolean strictMessageFilter) {
+        this.badWords = new ArrayList<String>();
+        this.memberOffenses = new HashMap<String, Integer>();
+        this.strictMessageFilter = strictMessageFilter;
 
         try {
             InputStream is = getClass().getClassLoader().getResourceAsStream("badwords.txt");
@@ -97,13 +99,14 @@ public class WordFilter extends ListenerAdapter {
                         """).queue();
                 }
                 else if (offenseCount >= 5) {
+                    if (!strictMessageFilter) return;
                     event.getGuild().addRoleToMember(
                         id,
                         event.getGuild().getRoleById(App.getMutedRoleId())).queue();
                     channel.sendMessageFormat("%s has been muted.", event.getAuthor().getAsMention());
                 }
 
-                message.delete().queue();
+                if (strictMessageFilter) message.delete().queue();
                 return;
             }
         }
